Keyframes for transformation properties.
Added the possibility to add/remove keyframes on transformation
properties + visual keyframe curve.

When adding or moving a keyframe, we don't use the matplotlib position
anymore. Instead, we compute the position the same way we do for the seek
logic, to make sure the playhead seeks precisely on the added/moved keyframe.
This change led to some other changes in the unit test which tested the
keyframe curve. More precisely, we needed to compute the position in pixels
for the click events, as that is how the seek logic received it.

Differential Revision: https://phabricator.freedesktop.org/D1766