clipproperties: Fixed a bug where keyframes were added on selection change
When changing the selected clip, a new keyframe was added incorrectly. This
happened because the transformation spin buttons values were updated
before clearing the previous control bindings, which was interpreted as
a new keyframe being added on the old selected clip. The fix consists of
updating the spin buttons values after clearing the previous control
bindings.

Differential Revision: https://phabricator.freedesktop.org/D1769